move interface, fix corner case
