+++ Assets/MixedRealityToolkit-SDK/Features/Audio/Influencers/AudioInfluencerController.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Microsoft.MixedReality.Toolkit.Core.Interfaces.Audio;
 using Microsoft.MixedReality.Toolkit.Core.Utilities;
 using System;
 using System.Collections.Generic;
@@ -169,7 +170,7 @@ namespace Microsoft.MixedReality.Toolkit.SDK.Audio.Influencers
             DateTime now = DateTime.Now;
 
             // Audio influences are not updated every frame.
-            if ((UpdateInterval * 1000.0f) <= (now - lastUpdate).Milliseconds)
+            if ((UpdateInterval * 1000.0f) <= (now - lastUpdate).TotalMilliseconds)
             {
                 audioSource.volume = initialAudioSourceVolume;
