+++ Assets/MixedRealityToolkit-SDK/Features/Audio/Influencers/AudioOccluder.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using Microsoft.MixedReality.Toolkit.Core.Extensions;
+using Microsoft.MixedReality.Toolkit.Core.Interfaces.Audio;
 using UnityEngine;
 
 namespace Microsoft.MixedReality.Toolkit.SDK.Audio.Influencers
