+++ Assets/MixedRealityToolkit/_Core/Interfaces/Audio.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a213c231b5994d24e8be6cea35786aaa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
